Move cocoapods cdn stubs to helper
