+++ cocoapods/spec/dependabot/cocoapods/file_updater_spec.rb
@@ -4,85 +4,14 @@ require "spec_helper"
 require "dependabot/dependency"
 require "dependabot/dependency_file"
 require "dependabot/cocoapods/file_updater"
+require_relative "helpers/cocoapods_cdn_stubs"
 require_common_spec "file_updaters/shared_examples_for_file_updaters"
 
 RSpec.describe Dependabot::CocoaPods::FileUpdater do
   it_behaves_like "a dependency file updater"
 
-  COCOAPODS_CDN_HOST = "https://cdn.cocoapods.org"
-
-  let(:cocoapods_all_pods) do
-    fixture("cocoapods", "all_pods", "all_pods.txt")
-  end
-
-  let(:cocoapods_deprecated_specs) do
-    fixture("cocoapods", "podspecs", "deprecated_podspecs.txt")
-  end
-
-  let(:cocoapods_version_yaml) do
-    fixture("cocoapods", "CocoaPods-version.yml")
-  end
-
-  let(:fixtures_path) do
-    File.join("spec", "fixtures", "cocoapods")
-  end
-
-  let(:all_versions_files) do
-    File.join(fixtures_path, "all_pods", "all_pods_versions_*.txt")
-  end
-
-  let(:podspec_files) do
-    File.join(fixtures_path, "podspecs", "*.podspec.json")
-  end
-
-  def stub_all_pods_versions_requests
-    Dir[all_versions_files].each do |file|
-      versions_url = "#{COCOAPODS_CDN_HOST}/#{File.basename(file)}"
-      stub_request(:get, versions_url).
-        to_return(status: 200, body: File.read(file))
-    end
-  end
-
-  def stub_all_spec_requests
-    spec_paths = {
-      "Nimble-2.0.0": "Specs/d/c/d/Nimble/2.0.0/Nimble.podspec.json",
-      "Nimble-3.0.0": "Specs/d/c/d/Nimble/3.0.0/Nimble.podspec.json",
-      "Alamofire-3.0.1": "Specs/d/a/2/Alamofire/3.0.1/Alamofire.podspec.json",
-      "Alamofire-3.5.1": "Specs/d/a/2/Alamofire/3.5.1/Alamofire.podspec.json",
-      "Alamofire-4.5.0": "Specs/d/a/2/Alamofire/4.5.0/Alamofire.podspec.json",
-      "Alamofire-4.5.1": "Specs/d/a/2/Alamofire/4.5.1/Alamofire.podspec.json",
-      "Alamofire-4.6.0": "Specs/d/a/2/Alamofire/4.6.0/Alamofire.podspec.json",
-      "AlamofireImage-2.5.0":
-          "Specs/8/0/a/AlamofireImage/2.5.0/AlamofireImage.podspec.json",
-      "AlamofireImage-4.1.0":
-          "Specs/8/0/a/AlamofireImage/4.1.0/AlamofireImage.podspec.json"
-    }
-
-    Dir[podspec_files].each do |file|
-      spec = File.basename(file, ".podspec.json")
-      spec_path = spec_paths[spec.to_sym]
-      versions_url = "#{COCOAPODS_CDN_HOST}/#{spec_path}"
-
-      stub_request(:get, versions_url).
-        to_return(status: 200, body: File.read(file))
-    end
-  end
-
   before do
-    all_pods_url = "#{COCOAPODS_CDN_HOST}/all_pods.txt"
-    stub_request(:get, all_pods_url).
-      to_return(status: 200, body: cocoapods_all_pods)
-
-    deprecated_specs_url = "#{COCOAPODS_CDN_HOST}/deprecated_podspecs.txt"
-    stub_request(:get, deprecated_specs_url).
-      to_return(status: 200, body: cocoapods_deprecated_specs)
-
-    cocoapods_version_url = "#{COCOAPODS_CDN_HOST}/CocoaPods-version.yml"
-    stub_request(:get, cocoapods_version_url).
-      to_return(status: 200, body: cocoapods_version_yaml)
-
-    stub_all_pods_versions_requests
-    stub_all_spec_requests
+    stub_all_cocoapods_cdn_requests
   end
 
   before do
